refactor: Transaction commands improvements (#350)
* Default account for consume-notes and refactors

* Further refactors

* Rename to utils

* Newline character

* CHANGELOG

* CHANGELOG

* Update src/cli/utils.rs


* Lints

---------
